Adding a compile-time setting on maximum OSTs for Lustre systems
A Lustre file system has a limited number of available OSTs to use.
For example, this number is 248 on Cori and 56 on Theta.

We should not go to full as Lustre will bypass assigning slow OSTs
to the file at file creation time, going to a recommended smaller
number (e.g. 72 on Cori and 48 on Theta) leaves room for this.

We also do not want to create high metadata requests and negatively
impact other users.

Set this limit for Cori and Theta so far with more machines that
use Lustre systems to be added later.